requirements.txt
flask
flask_session
flask_cors
requests
cv2
pillow
asyncio